requirements.txt
# NewsAPI Automation App Dependencies
requests==2.31.0
python-dotenv==1.0.0
replicate
Pillow
pyinstaller
